- added two 'mix' methods to determine range
The four range check methods are:
1) combatlog
2) class-spell based
3) mix 1: class-spell based + combatlog (range: 0-45)
4) mix 2: class-spell based + combatlog (range: class-spell dependent)

3/4) is not 1) + 2)
3/4) is 2) + combatlog (if you are attacked only)

3/4) covers one situation that 2) alone can't:
If you are attacked and no one in your group (including you) is targeting this attacker no range can be determined.
The included combatlog scan can cover such situations.

The CPU usage is the highest with method 3) or 4).
Disable the whole range check option if you have CPU/FPS problems in combat.